Create date_balance_summry.py
create a report with Year and Month as columns and the balance amount on 10', 20' and last day of the month as rows. If the balance on 7' of September was Rs. 120563.45, then the balance on 10' of September would be the same. If the balance for the required date i.e., 10', 20' and last day is not given, the previous balance will be desired balance.

 If the first month doesn't have the balance on or before date, the amount should be 0